:truck: Move: Move NavBar, Header
