--- src/components/Header.js
@@ -1,7 +1,0 @@
-import NavBar from './NavBar';
-
-const Header = () => {
-	return <NavBar />;
-};
-
-export default Header;
