+++ src/components/common/NavBar.js
@@ -1,23 +1,18 @@
 import styled from 'styled-components';
-import { ReactComponent as LogoSVG } from '../assets/icons/logo.svg';
-import { useEffect, useState } from 'react';
+import { ReactComponent as LogoSVG } from '../../assets/icons/logo.svg';
 
 const NavBarStyle = styled.nav`
-	position: sticky;
-	top: 0;
-	z-index: 10;
+	max-width: 1920px;
+	margin: 0 auto;
 	width: 100%;
 	height: 60px;
-	background-color: white;
-	border-bottom: ${(props) =>
-		`1px solid ${props.border ? 'rgba(0, 27, 55, 0.1)' : 'transparent'}}`};
 
 	.navbar-container {
 		display: flex;
 		justify-content: space-between;
 		align-items: center;
 		height: 100%;
-		margin: 0 24px;
+		padding: 0 24px;
 	}
 
 	.navbar-content {
@@ -39,23 +34,8 @@ const NavBarStyle = styled.nav`
 `;
 
 const NavBar = () => {
-	const [border, setBorder] = useState(false);
-
-	const onScrollHandler = () => {
-		if (window.scrollY > 0) setBorder((prev) => true);
-		else setBorder((prev) => false);
-	};
-
-	useEffect(() => {
-		window.addEventListener('scroll', onScrollHandler);
-
-		return () => {
-			window.removeEventListener('scroll', onScrollHandler);
-		};
-	}, []);
-
 	return (
-		<NavBarStyle border={border}>
+		<NavBarStyle>
 			<ul className="navbar-container">
 				<li className="navbar-content">
 					<span className="navbar-logo-container">
